agents/pluma/pluma-local/requirements.txt
# Google API and OAuth
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0

# Anthropic Claude API
anthropic>=0.7.0

# Redis for caching
redis>=5.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Rich terminal UI
rich>=13.5.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0

# Utilities
requests>=2.31.0
typing-extensions>=4.8.0